Doc change docCode from int to string

+++ src/main/java/ru/bellintegrator/practice/dao/impl/UserDAOImpl.java
@@ -116,7 +116,7 @@ public class UserDAOImpl implements UserDAO {
             objConjunction.add(Restrictions.like("possition", "%" + userFilterView.getPossition() + "%").ignoreCase());
         }
         if (userFilterView.getDocCode() != null) {
-            objConjunction.add((Restrictions.eq("doc.code", userFilterView.getDocCode().intValue())));
+            objConjunction.add((Restrictions.eq("doc.code", userFilterView.getDocCode())));
         }
         if (userFilterView.getCitizenShipCode() != null) {
             objConjunction.add(Restrictions.eq("country.code", userFilterView.citizenShipCode));
